+++ mayan/apps/documents/migrations/0050_auto_20190725_0451.py
@@ -4,7 +4,6 @@ from django.db import migrations, models
 
 
 class Migration(migrations.Migration):
-
     dependencies = [
         ('documents', '0049_auto_20190715_0454'),
     ]